[ghstack-poisoned]

+++ aten/src/ATen/functorch/TensorWrapper.cpp
@@ -195,7 +195,7 @@ static void dead_tensor_wrapper_fallback(const c10::OperatorHandle& op, torch::j
     return wrapped->value();
   };
 
-  foreachTensorInplace(*stack, stack->size() - args_size, stack->size(), unwrapIfDeadAndIncrement);
+  foreachTensorInplace(*stack, static_cast<int64_t>(stack->size() - args_size), static_cast<int64_t>(stack->size()), unwrapIfDeadAndIncrement);
   TORCH_INTERNAL_ASSERT(unwrapped_count > 0, "Should have at least one dead wrapper");
 
   // re-dispatch
